Standardise development scripts (#6)
* Remove restriction on which services to build/run

https://eaflood.atlassian.net/browse/FPD-284

This is mainly for consistency with other repositories where more than
one service is started from a single docker-compose command. Restricting
docker-compose commands to specified services is likely to trip up
developers if supporting services (e.g. a data store) are added to this
repository in future.

* Standardise scripts as setup, start and stop

* Tweak formatting of README

* Add script to run automated tests

* Add script for local Helm deployment

* Grammar correction in prerequisites

* Correct name of service in deploy script

* Add empty line under each heading in readme

* Rename setup script to build

This is more descriptive of its actual purpose as it will need to be run
whenever there are changes to the Dockerfile or dependencies.

* Pass build arguments to docker-compose

* Run tests through docker-compose

This makes the test script cleaner as the volume mounts are taken care
of in docker-compose.yaml.

* Move "How to run tests" to near top of readme

* Alphabetised scripts in package.json

* Remove unused "watch:build" task

This appears to have been erroneously copied from elsewhere as there is
nothing to build in this project.

* Run via nodemon in development

* Add postgres instance to Docker Compose

Run migrations on app container start.

* Correct name of postgres data volume

* Ignore .config/ directory generated by NPM

* Fix volume mounts for CI testing

Create Docker Compose override files for testing and local development
with appropriate volume mounts for each environment. This will enable
CI tests to run against code built into container image and output test
results to a bind-mounted folder on the host.

* Ignore .standard-v12-cache folder generated by NPM