Add healthcheck probes for Kubernetes (#15)
https://eaflood.atlassian.net/browse/FPD-471

The liveness probe confirms the service is running while the readiness probe tests database and message queue connections to ensure the service can properly serve requests. If the readiness probe fails, it responds with a list of disconnected services to assist in diagnosing problems.

Additionally, modify Helm scripts to include PostgreSQL and Artemis instances so that changes can be tested without connecting to remote services.